Greatly simplify the EnvVariables "window" logic.
Instead of building complicated EnvVariablesWindow which references an
original EnvVariables, allow the object to be cloned into a subset via
clone_into(). The content is couple strings anyways, so the overhead of
cloning is so small it, and it saves a lot of lifetime headaches. This
also neatly allows us to build windows from another window, since it's
just another EnvVariables object now.